requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
